# doc/openapi.yaml
openapi: "3.0.3"
info:
  title: AiR ORCHESTRATOR API
  description: |
    REST API сервиса AiR ORCHESTRATOR. Большинство защищённых маршрутов требуют
    заголовок `Authorization: Bearer <token>`. WebSocket-маршруты принимают
    токен через query-параметр `token` или через `Sec-WebSocket-Protocol`.
  version: "1.0.0"
servers:
  - url: https://localhost:8443
    description: Локальная разработка (TLS)
  - url: http://localhost:8080
    description: Production (за nginx)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AccessToken:
      type: object
      properties:
        s:
          type: string
          description: Новый STA-токен
...
paths:
...
  /auth/token/refresh:
    get:
      tags: [auth]
      summary: Обновление access-token на основе refresh-token из cookie
      responses:
        "200":
          description: Новый STA-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessToken"
        "401":
          description: Невалидный или отсутствующий refresh-token
...
  /dev/getdata:
    post:
      tags: [dev]
      summary: Получить текущие операционные настройки
      description: |
        Возвращает email рассылки, Google OAuth Client ID, имя Telegram-бота
        и дату создания текущего session key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                  google_id:
                    type: string
                  bot_name:
                    type: string
                  session_created:
                    type: string
                    format: date-time
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав (не администратор)
  /dev/setdistribmail:
    post:
      tags: [dev]
      summary: Установить SMTP-настройки рассылки
      description: |
        Сохраняет smtp.mail, smtp.host, smtp.port, smtp.pass в app_config (зашифровано).
        Пароль передаётся зашифрованным через сессионный ключ (AES-CBC).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, resp_id, mail, pass, host, port]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
                resp_id:
                  type: integer
                  format: uint64
                  description: RespId из POST /auth/session-key
                mail:
                  type: string
                  description: Email отправителя
                pass:
                  type: string
                  description: Зашифрованный пароль SMTP (AES-CBC)
                host:
                  type: string
                  description: SMTP-хост
                port:
                  type: string
                  description: SMTP-порт
      responses:
        "200":
          description: SMTP настройки сохранены
        "400":
          description: Ошибка парсинга или расшифровки пароля
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
  /dev/setgauth:
    post:
      tags: [dev]
      summary: Установить Google OAuth настройки
      description: |
        Сохраняет google_oauth.redirect_uri, google_oauth.client_id,
        google_oauth.client_secret в app_config (зашифровано).
        ClientSecret передаётся зашифрованным через сессионный ключ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, resp_id, url, id, sec]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
                resp_id:
                  type: integer
                  format: uint64
                  description: RespId из POST /auth/session-key
                url:
                  type: string
                  description: Google OAuth Redirect URI
                id:
                  type: string
                  description: Google Client ID
                sec:
                  type: string
                  description: Зашифрованный Google Client Secret (AES-CBC)
      responses:
        "200":
          description: Google OAuth настройки сохранены
        "400":
          description: Ошибка парсинга или расшифровки секрета
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
  /dev/setcarpintero:
    post:
      tags: [dev]
      summary: Установить настройки Telegram-бота (Carpintero)
      description: |
        Сохраняет tg.token, tg.bot, tg.id в app_config (зашифровано).
        Bot ID извлекается автоматически из токена (часть до первого `:`)
        BotToken передаётся зашифрованным через сессионный ключ.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, resp_id, bot-token, bot-name]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
                resp_id:
                  type: integer
                  format: uint64
                  description: RespId из POST /auth/session-key
                bot-token:
                  type: string
                  description: Зашифрованный токен Telegram-бота формата "ID:SECRET" (AES-CBC)
                bot-name:
                  type: string
                  description: Имя бота (без @)
      responses:
        "200":
          description: Telegram-бот настроен
        "400":
          description: Ошибка парсинга или расшифровки токена
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
  /dev/setsessionkey:
    post:
      tags: [dev]
      summary: Сгенерировать новый JWT session key
      description: |
        Сбрасывает auth.session и auth.created в app_config.
        При следующем запуске LoadOrInitKey генерирует новый JWT-ключ.
        Все существующие JWT становятся невалидными.
        auth.email_key НЕ затрагивается — хеши email пользователей сохраняются.
        После сброса приложение автоматически перезапускается.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
      responses:
        "200":
          description: Session key сброшен, приложение перезапускается
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
  /dev/checksettings:
    post:
      tags: [dev]
      summary: Проверить наличие ключевых конфигураций
      description: |
        Возвращает true/false для каждой группы настроек:
        smtp (smtp.mail), google_oauth (google_oauth.client_id), carpintero (tg.token).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
      responses:
        "200":
          description: Статус конфигураций
          content:
            application/json:
              schema:
                type: object
                properties:
                  smtp:
                    type: boolean
                    description: true — smtp.mail настроен
                  google_oauth:
                    type: boolean
                    description: true — google_oauth.client_id настроен
                  carpintero:
                    type: boolean
                    description: true — tg.token настроен
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
  /dev/getsvckey:
    post:
      tags: [dev]
      summary: Получить текущий gRPC service key
      description: |
        Возвращает расшифрованное значение svc.service_key из app_config.
        Используется если администратор потерял ключ и не хочет генерировать новый.
        Ключ предназначен для монтирования в микросервисы через Docker secrets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
      responses:
        "200":
          description: Текущий service key
          content:
            application/json:
              schema:
                type: object
                properties:
                  service_key:
                    type: string
                    description: Hex-строка сервисного ключа (64 символа)
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
        "404":
          description: Ключ не настроен — используйте POST /dev/generatesvckey
  /dev/generatesvckey:
    post:
      tags: [dev]
      summary: Сгенерировать новый gRPC service key
      description: |
        Генерирует новый случайный 32-байтный hex-ключ, сохраняет в app_config
        под ключом `svc.service_key` (зашифровано APP_MASTER_KEY).

        Возвращает plaintext ключ **один раз** — администратор должен сохранить его
        в файл `secrets/service_key.txt` каждого микросервиса.

        В `docker-compose.yml` микросервиса:
        ```
        SERVICE_KEY_FILE=/run/secrets/service_key.txt
        ```

        После генерации приложение автоматически перезапускается, чтобы gRPC-сервер
        стартовал с новым ключом. Микросервисы также необходимо перезапустить.

        Повторно получить plaintext можно через POST /dev/getsvckey.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: JWT access-token администратора
      responses:
        "200":
          description: Новый service key сгенерирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  service_key:
                    type: string
                    description: Plaintext hex-ключ (64 символа). Сохраните — более не будет показан в явном виде.
                    example: "a3f1...64hex...c9b2"
                  note:
                    type: string
                    description: Инструкция по сохранению ключа
        "401":
          description: Невалидный токен
        "403":
          description: Недостаточно прав
        "500":
          description: Ошибка генерации или сохранения ключа

  # ─────────────── gRPC ConfigService ───────────────
  # Описание gRPC-сервиса (не HTTP): порт задаётся через GRPC_PORT (по умолчанию 50051).
  # Транспорт: TLS. Авторизация: метаданные "x-service-key" = SERVICE_KEY_FILE.
  # Proto-файл: internal/delivery/grpc/proto/svc_config.proto
  #
  # rpc GetBotConfig(GetBotConfigRequest) returns (BotConfigResponse)
  #   — возвращает расшифрованные tg.token, tg.bot, tg.id из app_config.
  #   Ключ активируется через POST /dev/generatesvckey.